added auto-scroll function
added an auto scroll function with speed changeable by the user. the auto scroll will scroll from the top to the bottom of the page and then do the same from the bottom to the top, jquery was needed to make it work, so i also added some lines to implement it in every site the user will visit.